Separate hook tests in the separate test
