+++ tests/CustomPostType/CustomPostTypeExampleTest.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Tests\Unit\CustomPostType;
+
+use Brain\Monkey;
+use EightshiftBoilerplate\CustomPostType\PostTypeExample;
+
+/**
+ * Mock before tests.
+ */
+beforeEach(function () {
+	Monkey\setUp();
+});
+
+/**
+ * Cleanup after tests.
+ */
+afterEach(function () {
+	Monkey\tearDown();
+});
+
+
+test('Register method will call init hook', function () {
+	(new PostTypeExample())->register();
+
+	$this->assertSame(10, has_action('init', 'EightshiftBoilerplate\CustomPostType\PostTypeExample->postTypeRegisterCallback()'));
+});
